Remove healthy.php and related references from configuration and scripts

+++ config/start.sh
@@ -27,14 +27,18 @@ if [ ! -d /storage/php ]; then
     ln -s /storage/php/php.ini /storage/php/8.2/fpm/php.ini >> /storage/app.log
     ln -s /storage/php/php.ini /storage/php/8.2/cli/php.ini >> /storage/app.log
 fi
-# rm -vr /var/setup >> /storage/app.log
+if [ -d /var/setup ]; then
+    rm -vr /var/setup >> /storage/app.log
+fi
 
-echo "--- Symlink to /storage ---"
-ln -s /storage/.env /app/ >> /storage/app.log
-ln -s /storage/fstab /etc/ >> /storage/app.log
-ln -s /storage/nginx /etc/ >> /storage/app.log
-ln -s /storage/php /etc/ >> /storage/app.log
-ln -s /storage/www / >> /storage/app.log
+if [ ! -f /storage/.env ]; then
+    echo "--- Symlink to /storage ---"
+    ln -s /storage/.env /app/ >> /storage/app.log
+    ln -s /storage/fstab /etc/ >> /storage/app.log
+    ln -s /storage/nginx /etc/ >> /storage/app.log
+    ln -s /storage/php /etc/ >> /storage/app.log
+    ln -s /storage/www / >> /storage/app.log
+fi
 if [ ! -L /etc/letsencrypt ]; then
     ln -s /storage/webconfig/ssl /etc/letsencrypt >> /storage/app.log
 fi
@@ -47,7 +51,6 @@ if [ ! -f /www/default/index.html ]; then
     echo "--- Generate default /www ---"
     mkdir -p /www/default/
     cp -v /storage/webconfig/index.html /www/default/  >> /storage/app.log
-    cp -v /storage/webconfig/healty.php /www/default/  >> /storage/app.log
     chown -vR apps:apps /www/ >> /storage/app.log
 fi
 
@@ -67,7 +70,7 @@ if [ ! -f /storage/db.sqlite ]; then
     echo "--- Install Database (SQLite) ---"
     touch /storage/db.sqlite
     cd /app && artisan migrate --force >> /storage/app.log
-    cd /app && artisan db:seed >> /storage/app.log
+    cd /app && artisan db:seed --force >> /storage/app.log
 fi
 
 if [ ! -f /storage/webconfig/ssl/live/default/cert.pem ]; then
